+++ fabric/src/main/java/org/thinkingstudio/fork/nightconfigfixes/NightConfigFixes.java
@@ -1,0 +1,8 @@
+package org.thinkingstudio.fork.nightconfigfixes;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class NightConfigFixes {
+    public static final Logger LOGGER = LoggerFactory.getLogger(NightConfigFixes.class);
+}