+++ fabric/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Changelog
 - update dependencies
-- (Fabric) fix Indium incompatibility (#31)
+- (Fabric) try fix nightconfig issue